Average Score Calculator
The problem is about the user inputting the amount of students and subjects. Based on those numbers, each student will have as many inputs as the number of subjects. The inputs should be their score on a subject. The subjects are not specific, just called "score". After that, the average of each student is needed and summing them all up and finding the average of that based on the number of students. The code does exactly that.